+++ bin/node/src/commands/bundled.rs
@@ -315,6 +315,7 @@ impl BundledCommand {
                 store_ntx_builder_url,
                 block_producer_url,
                 validator_url,
+                &data_directory,
             );
 
             let id = join_set